fix : notification 분리 및 useProducts분리
