+++ src/basic/models/products/types.ts
